requirements.txt
numpy>=1.23
opencv-python>=4.7
scikit-image>=0.19
Pillow>=9.0
matplotlib>=3.5
torch>=1.12
torchvision>=0.13